fix: StudyDetailService.js EmojiService.js 로 변경하면서 그 안에 있던 Study, Habit 정보 알맞는 .js 에 옮김
