+++ src/pages/StudyDetailPage/StudyDetailPage.jsx
@@ -9,7 +9,7 @@ import Modals from './components/Modals/Modals';
 
 import Toast from '../../components/Toast/Toast';
 
-import { getStudyDetail } from '../../services/StudyDetailService.js';
+import { getStudyDetail } from '../../services/StudyService.js';
 
 import styles from './StudyDetailPage.module.css';
